@@ -9,11 +9,13 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=0.10.0"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
-		"test": "ava"
+		"test": "xo && ava"
 	},
 	"files": [
 		"index.js"
@@ -26,10 +28,13 @@
 		"limit",
 		"min",
 		"max",
+		"minimum",
+		"maximum",
 		"between",
 		"range"
 	],
 	"devDependencies": {
-		"ava": "^0.25.0"
+		"ava": "^3.15.0",
+		"xo": "^0.44.0"
 	}
 }
